Paypal Integration

+++ com_events/site/controllers/unregister.php
@@ -17,18 +17,35 @@
 			
 			$return = array("success" => false);
 			
-			$model = new EventsModelsEvent();
+			$model = null
 			$eventView = null;
 			
-			if($model->deleteAttendee())
+			// Gets current view.
+			$view = $app->input->get('view', 'event');
+			
+			// if calling from the event view.
+			if($view == 'event')
 			{
-				$return['success'] = true;
-				$eventView = EventsHelpersView::load('event','_result-unregister-success','phtml');
-			}
-			else
-			{
-				$return['success'] = false;
-				$eventView = EventsHelpersView::load('event','_result-unregister-failure','phtml');
+				$model = new EventsModelsEvent();
+				
+				// Gets the current user that is logged in
+				$currentUser = $model->getCurrentUser();
+				
+				// If the user has signed up for the event and isn't paid then allow it to be removed.
+				if(isset($currentUser->status) && ((int) $currentUser->status == 2 || (int) $currentUser->status == 1)) 
+				{
+					if($model->deleteAttendee())
+					{
+						$return['success'] = true;
+						$eventView = EventsHelpersView::load('event','_result-unregister-success','phtml');
+					}
+					else
+					{
+						$return['success'] = false;
+						$eventView = EventsHelpersView::load('event','_result-unregister-failure','phtml');
+					}
+					
+				}	
 			}
 			ob_start();
 			echo $eventView->render();
@@ -36,7 +53,7 @@
 			ob_clean();
 			 
 			$return['html'] = $html;
-				
+			
 			echo json_encode($return);
 		}
 	}
